+++ src/services/product.js
@@ -1,24 +1,24 @@
 const product = require("../model/product");
 
 class ServiceProduct {
-    async FindAll(transaction) {
-        return product.findAll({ transaction });
+    async FindAll(organizationId, transaction) {
+        return product.findAll({ where: { organizationId }, transaction });
     }
 
-    async FindById(id, transaction) {
-        return product.findByPk(id, { transaction });
+    async FindById(organizationId, id, transaction) {
+        return product.findOne({ where: { id, organizationId }, transaction });
     }
 
     async Create(name, description, organizationId, transaction) {
         if (!name) {
-            throw new Error("Please provide product name");
+            throw new Error("Favor informar o nome do produto");
         }
 
         return product.create({ name, description, organizationId }, { transaction });
     }
 
-    async Update(id, name, description, transaction) {
-        const existingProduct = await this.FindById(id, transaction);
+    async Update(organizationId, id, name, description, transaction) {
+        const existingProduct = await this.FindById(organizationId, id, transaction);
 
         existingProduct.name = name || existingProduct.name;
         existingProduct.description = description || existingProduct.description;
@@ -28,8 +28,13 @@ class ServiceProduct {
         return existingProduct;
     }
 
-    async Delete(id, transaction) {
-        const existingProduct = await this.FindById(id, transaction);
+    async Delete(organizationId, id, transaction) {
+        const existingProduct = await this.FindById(organizationId, id, transaction);
+
+        if(!existingProduct) {
+            throw new Error("Favor informar o produto corretamente")
+        }
+
         await existingProduct.destroy({ transaction });
         return true;
     }
